+++ logger.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const logger = require('./lib/logger');
+
+exports.setLoggerProvider = logger.setLoggerProvider;