+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Strip uknown fields in client requests.
 - Disable x-powered-by header
 - Improved client-metrics validation to avoid NaN
+- Add posibility to inject custom logger provider
 
 ## 3.0.0-alpha.1
 - upgrade unleash-frontend to 3.0.0-alpha.1